trainers/toy_model/requirements.txt
scikit-learn==1.1.3
numpy==1.23.4
matplotlib==3.6.2
pandas==1.5.1
seaborn==0.12.1
fsspec==2022.11.0
gcsfs==2022.11.0
google-cloud-storage==2.7.0
gcloud==0.18.3